Save current work on option handle refactor
Instead of having Raku parse all command line arguments into a list
(for the positionals) and a hash (for the named options), we're going
to go through @*ARGS, look at the name and execute a subroutine by
that name (prefixed with (option-) to handle that flag.

Only one action (out of 14 at current count) can be done at the same
time.  Look at the name of the action and execute a subroutine by
that name (prefixed with action-) to do that action.

This is by no means finished and won't even compile.  But it represents
several days of work, so I thought of putting it in a safe place for
now.