Massive refactor of App::Rak
Almost from scratch, specifically with regards to the way the
command line arguments are handled.  Instead of using the standard
MAIN idiom (in which named arguments are placed into a hash and
then slipping them into the call to &MAIN), this approach just
runs over @*ARGS and if there's a named argument (something starting
with "-"), attempts to call the associated "option-$name" subroutine
which is then responsible for setting the right data structures for
that named argument.  Some of these actually set the "action" to
be performed.

Once all arguments (and their expansions from the user configuration)
have been parsed, there should be an action to be performed.  Of there
is no action, then "per-line" is assumed.  Then the sub "action-$name"
is then called to actually perform the expected action.

This approach makes it easier to add new flags (add an "option-$name"
subroutine) and new actions (have the "option-$name" sub set an
action name, and make sure there is an "action-$name" sub).

This approach should also make it easier to provide better feecback
in combinations of arguments that are not allowed, because they do
not make any sense.  For instance, setting highlighting options
if you're only interested in editing the location of the search
result.

This commit is just to make sure all this work is saved.  The current
state of this code, is that it compiles and that it doesn't die
when 'use'ing the module.  Specifically the highlighting logic needs
to be refactored in still.  At least this compiles  :-)