Remove the --find-all option
When the default was --known-extensions, it made sense.  Now that the
default is --is-text, --find-all has basically become a synonym for a
bare --dir (as the default for --file is bare as well).

This is the first time an option has been *removed*.  I don't expect
this to be happening much anymore.  --find-all always felt like a
bit of a hack anyway.